+++ app/src/main/java/com/bgcoding/notes/app/feature_note/presentation/notes/NotesScreen.kt
@@ -21,6 +21,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
@@ -29,11 +30,14 @@ import androidx.compose.material.icons.automirrored.filled.Sort
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.CameraAlt
 import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.DeleteForever
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.Notes
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material.icons.filled.Sort
+import androidx.compose.material3.AlertDialogDefaults
+import androidx.compose.material3.BasicAlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -50,15 +54,21 @@ import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SnackbarResult
+import androidx.compose.material3.Surface
 import androidx.compose.material3.SwipeToDismissBox
 import androidx.compose.material3.SwipeToDismissBoxValue
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.material3.rememberSwipeToDismissBoxState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.rememberUpdatedState
@@ -70,6 +80,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
+import com.bgcoding.notes.app.feature_note.presentation.add_edit_note.AddEditNoteEvent
 import com.bgcoding.notes.app.feature_note.presentation.notes.components.NoteItem
 import com.bgcoding.notes.app.feature_note.presentation.notes.components.OrderSection
 import com.bgcoding.notes.app.feature_note.presentation.util.Screen
@@ -79,12 +90,87 @@ import kotlinx.coroutines.launch
 @Composable
 fun NotesScreen(
     navController: NavController,
-    viewModel: NotesViewModel = hiltViewModel()
+    viewModel: NotesViewModel = hiltViewModel(),
 ) {
+
+    // get showDeleted from navController argument
+    val showDeleted = navController.currentBackStackEntry?.arguments?.getBoolean("showDeleted") ?: false
+
+    LaunchedEffect(showDeleted) {
+        viewModel.onEvent(NotesEvent.SetShowDeleted(showDeleted))
+        Log.d("NotesScreen", "in LaunchedEffect showDeleted: $showDeleted")
+    }
+
     val state = viewModel.state.value
     val snackbarHostState = remember { SnackbarHostState() }
     val drawerState = rememberDrawerState(DrawerValue.Closed)
     val scope = rememberCoroutineScope()    // scoped to the composable
+
+    val showDialog = remember { mutableStateOf(false) }
+
+    if (showDialog.value) {
+        BasicAlertDialog(
+            onDismissRequest = { showDialog.value = false }
+        ) {
+            Surface(
+                modifier = Modifier
+                    .padding(16.dp),
+                shape = MaterialTheme.shapes.large,
+                tonalElevation = AlertDialogDefaults.TonalElevation
+            ) {
+                Column(
+                    modifier = Modifier.wrapContentSize()
+                ) {
+                    Text(
+                        text = "Delete Notes",
+                        style = MaterialTheme.typography.titleLarge,
+                        modifier = Modifier.padding(16.dp).align(Alignment.CenterHorizontally)
+                    )
+                    Text(
+                        text = "Are you sure you want to permanently delete all notes?",
+                        style = MaterialTheme.typography.bodyMedium,
+                        modifier = Modifier.padding(horizontal = 16.dp).align(Alignment.CenterHorizontally)
+                    )
+                    Spacer(modifier = Modifier.height(4.dp))
+                    Row(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(
+                                start = 16.dp,
+                                end = 16.dp,
+                                bottom = 16.dp
+                            ),
+                        horizontalArrangement = Arrangement.End
+                    ) {
+                        TextButton(
+                            onClick = {
+                                showDialog.value = false
+                            }
+                        ) {
+                            Text(
+                                "Cancel",
+                                style = MaterialTheme.typography.bodyMedium,
+                                color = MaterialTheme.colorScheme.error
+                            )
+                        }
+                        Spacer(modifier = Modifier.width(8.dp))
+                        TextButton(
+                            onClick = {
+                                showDialog.value = false
+                                viewModel.onEvent(NotesEvent.DeleteAllNotesPermanently)
+                            }
+                        ) {
+                            Text(
+                                "Delete",
+                                style = MaterialTheme.typography.bodyMedium,
+                                color = MaterialTheme.colorScheme.error
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
 
     // Log recompositions
     val currentState = rememberUpdatedState(state)
@@ -92,6 +178,10 @@ fun NotesScreen(
         Log.d("NotesScreen", "Recomposed with state: $currentState")
         onDispose { }
     }
+
+
+
+    Log.d("NotesScreen", "showDeleted: $showDeleted")
 
     ModalNavigationDrawer(
         drawerState = drawerState,
@@ -122,8 +212,10 @@ fun NotesScreen(
                     label = {
                         Text(text = "Notes", style = MaterialTheme.typography.bodyMedium, fontWeight = FontWeight.Bold)
                     },
-                    selected = true,
-                    onClick = { /*TODO*/ }
+                    selected = !showDeleted,
+                    onClick = {
+                        navController.navigate(Screen.NotesScreen.route + "?showDeleted=false")
+                    }
                 )
                 NavigationDrawerItem(
                     icon = {
@@ -135,8 +227,10 @@ fun NotesScreen(
                     label = {
                         Text(text = "Bin", style = MaterialTheme.typography.bodyMedium, fontWeight = FontWeight.Bold)
                     },
-                    selected = false,
-                    onClick = { /*TODO*/ }
+                    selected = showDeleted,
+                    onClick = {
+                        navController.navigate(Screen.NotesScreen.route + "?showDeleted=true")
+                    }
                 )
                 NavigationDrawerItem(
                     icon = {
@@ -204,12 +298,21 @@ fun NotesScreen(
                 )
             },
             floatingActionButton = {
-                FloatingActionButton(onClick = {
-                    navController.navigate(Screen.AddEditNoteScreen.route)
+                FloatingActionButton(onClick =
+                {
+                    if (!showDeleted) {
+                        navController.navigate(Screen.AddEditNoteScreen.route)
+                    } else {
+                        showDialog.value = true
+                    }
                 },
                     containerColor = MaterialTheme.colorScheme.primaryContainer,
                 ) {
-                    Icon(imageVector = Icons.Filled.Add, contentDescription = "Add note")
+                    if (!showDeleted)
+                        Icon(imageVector = Icons.Filled.Add, contentDescription = "Add note")
+                    else {
+                        Icon(imageVector = Icons.Filled.DeleteForever, contentDescription = "Delete Notes")
+                    }
                 }
             },
             snackbarHost = { SnackbarHost(snackbarHostState) },
@@ -238,13 +341,13 @@ fun NotesScreen(
                     .background(MaterialTheme.colorScheme.surface)
                 ) {
                     items(state.notes, key = { it.id!! }) { note ->
-                        /*DisposableEffect(Unit) {
+                        DisposableEffect(Unit) {
                             Log.d("LazyColumn", "LazyColumn recomposed")
                             onDispose {}
-                        }*/
+                        }
                         val dismissState = rememberSwipeToDismissBoxState(
                             confirmValueChange = {
-                                if (it == SwipeToDismissBoxValue.EndToStart) {
+                                if (!showDeleted && it == SwipeToDismissBoxValue.EndToStart) {
                                     viewModel.onEvent(NotesEvent.DeleteNote(note))
                                     scope.launch {
                                         val result = snackbarHostState
@@ -297,7 +400,6 @@ fun NotesScreen(
                                         navController.navigate(
                                             Screen.AddEditNoteScreen.route +
                                                     "?noteId=${note.id}"
-//                                        "?noteId=${note.id}&noteColor=${note.color}"
                                         )
                                     }
                             )
